feat: 구현 코드 변경에 따른 테스트 코드 변경

+++ src/test/java/lotto/model/LottoGameTest.java
@@ -47,7 +47,7 @@ public class LottoGameTest {
         Map<DrawType, Integer> drawResult = lottoGame.generateDrawResult();
 
         // when
-        long result = lottoGame.calculateEarns(drawResult, 5000);
+        double result = lottoGame.calculateEarns(drawResult, 5000);
 
         // then
         assertThat(result).isEqualTo(40631100);
